server/pyproject.toml
[tool.poetry]
name = "prompton"
version = "0.0.1"
description = ""
authors = ["szerintedmi <szerintedmi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.0"
motor = "^3.1.2"
uvicorn = "^0.22.0"
httpx = "^0.24.1"
aiohttp = "^3.8.4"
openai = "^0.27.8"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-multipart = "^0.0.6"
pydantic = { version = "^1.10.8", extras = ["email", "dotenv"] }
mangum = "^0.17.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
pytest = "^7.3.1"
mongomock-motor = "^0.0.19"
deepdiff = "^6.3.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
types-pyyaml = "^6.0.12.10"
types-python-jose = "^3.3.4.7"
types-passlib = "^1.7.7.12"
pyright = "^1.1.313"
flake8 = "^6.0.0"
flake8-black = "^0.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
